feat(uring): don't wait twice if looping in _get_cq_entry().
refactor(uring): use struct instead of tuple to return from __peek_cq_entry().

perf(uring): use template bool to indicate if timeout is needed

+++ include/uring/syscall.hpp
@@ -20,8 +20,8 @@ static inline void *ERR_PTR(intptr_t n) {
     return (void *)n;
 }
 
-static inline intptr_t PTR_ERR(const void *ptr) {
-    return (intptr_t)ptr;
+static inline int PTR_ERR(const void *ptr) {
+    return (int)(intptr_t)ptr;
 }
 
 static inline bool IS_ERR(const void *ptr) {
